requirements.txt
# Core dependencies
streamlit>=1.28.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1

# Multimodal processing
python-docx>=0.8.11
speechrecognition>=3.10.0
pydub>=0.25.1
librosa>=0.10.0
openai-whisper>=20231117
opencv-python>=4.8.0
pytesseract>=0.3.10
faiss-cpu>=1.7.4
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0

# PDF processing
unstructured>=0.10.0
pdfminer.six==20231228
pymupdf>=1.23.0
pdf2image>=1.17.0

# Image processing  
pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10

# Data handling
pandas>=2.0.0
numpy>=1.24.0
loguru>=0.7.0
tqdm>=4.65.0

# Additional utilities
python-dotenv>=1.0.0
typing-extensions>=4.8.0
plotly>=5.17.0

# Optional dependencies for advanced features
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development and testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.0.0
mypy>=1.5.0

# Production deployment
gunicorn>=21.2.0
uvicorn>=0.23.0
